remove circular dependency in treeView
